chore: move checkLevelDbVersion to src/storage/leveldb/utils.ts
